@@ -4,15 +4,15 @@
   "private": true,
   "dependencies": {
     "@apollo/client": "^3.7.10",
-    "react": "18.2.0",
-    "react-dom": "18.2.0"
+    "react": "18.3.1",
+    "react-dom": "18.3.1"
   },
   "devDependencies": {
     "@graphql-codegen/client-preset-swc-plugin": "0.2.0",
-    "@graphql-codegen/cli": "^5.0.0",
+    "@graphql-codegen/cli": "^5.0.2",
     "@vitejs/plugin-react-swc": "^3.3.0",
-    "@types/react": "18.2.21",
-    "@types/react-dom": "18.2.7",
+    "@types/react": "18.3.1",
+    "@types/react-dom": "18.3.0",
     "typescript": "5.2.2",
     "vite": "^4.1.0"
   },
