@@ -43,17 +43,17 @@
     "examples/**/*"
   ],
   "devDependencies": {
-    "@babel/core": "7.22.11",
-    "@babel/preset-env": "7.22.10",
-    "@babel/preset-typescript": "7.22.11",
-    "@changesets/changelog-github": "0.4.8",
-    "@changesets/cli": "2.26.2",
-    "@theguild/eslint-config": "0.11.0",
+    "@babel/core": "7.24.5",
+    "@babel/preset-env": "7.24.5",
+    "@babel/preset-typescript": "7.24.1",
+    "@changesets/changelog-github": "0.5.0",
+    "@changesets/cli": "2.27.1",
+    "@theguild/eslint-config": "0.11.8",
     "@theguild/prettier-config": "0.1.1",
     "@types/jest": "28.1.8",
     "babel-jest": "29.6.4",
     "bob-the-bundler": "6.0.0",
-    "eslint": "8.47.0",
+    "eslint": "8.57.0",
     "graphql": "16.8.0",
     "husky": "8.0.3",
     "jest": "28.1.3",
@@ -63,11 +63,11 @@
     "patch-package": "8.0.0",
     "prettier": "2.8.8",
     "prettier-plugin-svelte": "2.10.1",
-    "rimraf": "4.4.1",
+    "rimraf": "5.0.5",
     "ts-jest": "28.0.8",
     "ts-node": "10.9.1",
     "tslib": "2.6.2",
-    "tsx": "3.12.7",
+    "tsx": "4.9.3",
     "typescript": "5.2.2"
   },
   "lint-staged": {
