@@ -66,16 +65,15 @@
     "@graphql-codegen/typescript-vue-apollo-smart-ops": "2.3.6",
     "@graphql-codegen/typescript-vue-urql": "2.3.6",
     "@graphql-codegen/urql-introspection": "2.2.1",
-    "@mendable/search": "0.0.155",
     "@mizdra/graphql-codegen-typescript-fabbrica": "^0.3.0",
     "@monaco-editor/react": "4.5.2",
-    "@theguild/components": "^6.0.1",
+    "@theguild/components": "6.5.3",
     "classnames": "2.3.2",
-    "date-fns": "2.29.3",
+    "date-fns": "2.30.0",
     "dedent": "1.5.1",
     "graphql": "16.8.0",
     "js-yaml": "4.1.0",
-    "next": "^13.5.4",
+    "next": "14.2.3",
     "next-mdx-remote": "4.4.1",
     "next-sitemap": "4.2.2",
     "react": "^18.2.0",
