@@ -1,7 +1,7 @@
 import os from 'os';
 import { join, parse, relative, resolve } from 'path';
 import makeDir from 'make-dir';
-import rimraf from 'rimraf';
+import * as rimraf from 'rimraf';
 
 const fs = jest.requireActual('fs');
 
