@@ -12,11 +12,10 @@ import { TypeScriptResolversVisitor } from './visitor.js';
 
 const capitalize = (s: string): string => s.charAt(0).toUpperCase() + s.slice(1);
 
-export const plugin: PluginFunction<TypeScriptResolversPluginConfig, Types.ComplexPluginOutput> = (
-  schema: GraphQLSchema,
-  documents: Types.DocumentFile[],
-  config: TypeScriptResolversPluginConfig
-) => {
+export const plugin: PluginFunction<
+  TypeScriptResolversPluginConfig,
+  Types.ComplexPluginOutput<{ generatedResolverTypes: Record<string, { name: string }> }>
+> = (schema: GraphQLSchema, documents: Types.DocumentFile[], config: TypeScriptResolversPluginConfig) => {
   const imports = [];
   if (!config.customResolveInfo) {
     imports.push('GraphQLResolveInfo');
@@ -280,6 +279,8 @@ export type DirectiveResolverFn<TResult = {}, TParent = {}, TContext = {}, TArgs
 
   prepend.push(...mappersImports, ...visitor.globalDeclarations);
 
+  const rootResolver = getRootResolver();
+
   return {
     prepend,
     content: [
@@ -289,9 +290,12 @@ export type DirectiveResolverFn<TResult = {}, TParent = {}, TContext = {}, TArgs
       resolversTypeMapping,
       resolversParentTypeMapping,
       ...visitorResult.definitions.filter(d => typeof d === 'string'),
-      getRootResolver(),
+      rootResolver.content,
       getAllDirectiveResolvers(),
     ].join('\n'),
+    meta: {
+      generatedResolverTypes: rootResolver.generatedResolverTypes,
+    },
   };
 };
 
