@@ -1,15 +1,28 @@
 /* eslint-disable import/no-extraneous-dependencies, no-console */
-// @ts-check
 const fs = require('fs-extra');
 const fg = require('fast-glob');
 
 const packageJSON = fg.sync(['examples/**/package.json'], { ignore: ['**/node_modules/**'] });
 
-console.log(
-  packageJSON
-    .map(packageJSONPath => {
-      const { name } = fs.readJSONSync(packageJSONPath);
-      return `yarn workspace ${name} run ${process.argv[2]}`;
-    })
-    .join(' && ')
+const ignoredPackages = [];
+
+const result = packageJSON.reduce(
+  (res, packageJSONPath) => {
+    const { name } = fs.readJSONSync(packageJSONPath);
+
+    if (ignoredPackages.includes(name)) {
+      res.ignored.push(name);
+      return res;
+    }
+
+    res.commands.push(`yarn workspace ${name} run ${process.argv[2]}`);
+    return res;
+  },
+  { ignored: [], commands: [] }
 );
+
+if (result.ignored.length > 0) {
+  result.commands.push(`echo "Ignored packages: ${result.ignored.join(',')}"`);
+}
+
+console.log(result.commands.join(' && '));
